Create wandb sync script
We found that simply copying all logs and syncing is not enough since
wandb re-uploads every run, making it slower as we train more models.

This script keeps a synced version of the logs locally. Then it
copies the younger jobs to a new folder and syncs only these, which is
much faster.