Implement chunk splitting in the dataloader
Instead of loading variable length samples with batch size 1, the
dataloader generates tuples of (chunk, filename). Then the predict
method stores the batch data into a dict (filename [chunk 0, chunk 1...]
and the embedding data is aggregated and stored to file at the epoch
end.

Note that this approach allows for much higher parallelization.