Do not squeeze the batch axis of EnCodec
We want reps to be (B, D, T).
This is evaluated in maskingmodel.py in line 451:

```
assert x.shape[1] == self.patch_size[0]
```

When batch_size=1, this assertion fails, on encoded because of the
squeeze operator. Thus, let's remove it.

Maybe this was not an issue before since we never used encodec
models with batch size 1.